Move the search page filter to a local folder.
