+++ meteor-app/imports/ui/pages/search-by-map/filter-components/data-temporal-range-filter.js
